packages/nex-staging/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nex-staging"
version = "0.1.0"
description = "Shared PostgreSQL models and repositories for invoice staging"
requires-python = ">=3.9"
dependencies = [
    "psycopg2-binary>=2.9.0",
    "pydantic>=2.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["nex_staging"]
